chore(seed): seed EventRSVP rows for roster/waitlist/stats QA (Issue 563)
Extend the seed command to create EventRSVP rows so a developer running
`make seed` gets a realistic, QA-able RSVP state:

- All four statuses (attending/maybe/cant_go/waitlisted) across events.
- "Vegan Potluck" filled to max_attendees=3 so the full/waitlist UI shows.
- "Past Potluck (seed)" has attended/no_show marked so the host stats
  panel renders populated.
- Some RSVPs set has_plus_one=True.
- Rows written directly via get_or_create so seeded statuses (incl.
  waitlisted) land verbatim and the seed stays idempotent.

Also extract the static seed data (dataclasses + constants) out of
seed.py into a sibling _seed_data.py module for file-size hygiene.
